requirements.txt
numpy>=1.24
matplotlib>=3.7
scipy>=1.10
h5py>=3.8
requests>=2.28
Pillow>=9.5
tqdm>=4.65
astropy>=5.3

# For HPC / local mode only:
# illustris_python>=1.0  (pip install illustris_python)
